+++ src/compas_timber/elements/fasteners/ball_node_fastener.py
@@ -7,8 +7,7 @@ from compas.geometry import Point
 from compas.geometry import Sphere
 from compas.geometry import Transformation
 from compas.geometry import Vector
-from compas.geometry import angle_vectors_signed
-
+from compas_timber.utils import correct_polyline_direction
 from compas_timber.elements import CutFeature
 from compas_timber.elements import Fastener
 from compas_timber.elements import FastenerTimberInterface
@@ -156,27 +155,3 @@ class BallNodeFastener(Fastener):
                     self._interface_shape += geometry
         return self._interface_shape
 
-
-def correct_polyline_direction(polyline, normal_vector):
-    """Corrects the direction of a polyline to be counter-clockwise around a given vector.
-
-    Parameters
-    ----------
-    polyline : :class:`compas.geometry.Polyline`
-        The polyline to correct.
-
-    Returns
-    -------
-    :class:`compas.geometry.Polyline`
-        The corrected polyline.
-
-    """
-    angle_sum = 0
-    for i in range(len(polyline) - 1):
-        u = Vector.from_start_end(polyline[i - 1], polyline[i])
-        v = Vector.from_start_end(polyline[i], polyline[i + 1])
-        angle = angle_vectors_signed(u, v, normal_vector)
-        angle_sum += angle
-    if angle_sum > 0:
-        polyline = polyline[::-1]
-    return polyline
